backend/requirements.txt
# OMEGA Core v3.0 - Python Dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
asyncpg==0.29.0

# Authentication & Security
PyJWT==2.9.0
cryptography==43.0.1
bcrypt==4.2.0
passlib[bcrypt]==1.7.4

# Rate Limiting
slowapi==0.1.9

# HTTP Client
httpx==0.27.2

# AI Services
anthropic==0.39.0
openai==1.51.0

# Google APIs
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.150.0

# Payment Processing
stripe==7.8.0

# SMS Integration
twilio==8.10.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.11.2

